Upgrade to latest (at the time of writing) version of Caliper
This does not upgrade to a released version of Caliper, it's
just a snapshot, see README.android for details.

Since 0.5-rc1 Caliper has gone through a huge number of changes
but it still does essentially the same job in the same way it
is just more flexible and capable than the previous version.

One major change that impacts the build is that this version
uses the Dagger2 dependency injection framework that generates
some of its code using annotation processors and so it was
necessary to change the build to use Dagger2.

Not all the Caliper tests run on Android, some are written for
the JVM and rely on some characteristic of it that is not the
same on Android, whether it be available memory, exception
message, supported classes or file structure. There are also a
couple of failures that are not completely explainable at the
moment. The details of these failures are listed in the
expectations/knownfailures.txt and will be addressed at a later
date.

Bug: 24848946
Change-Id: Ia245db56c57315ce18db8eb219003fecf3c64ab9