Update caliper for guava 27.1
The current version of caliper is not compatible with guava 27.1.  In
order to decouple updating guava from updating caliper, this CL does
the minimal changes to caliper to compile with guava 27.1.  It should
ideally be reverted and replaced with merging an upstream version of
caliper that supports guava 27.1.

Use java 1.8 to get the new default method in the Predicate interface.
@Provides(type = Type.SET) has been replaced with @Provides @IntoSet.

Futures.withFallback has been replaced with Futures.catchingAsync.

Error Prone now requires checking the return value of
ListeningExecutorService.submit.

Bug: 130306229
Test: m checkbuild
Change-Id: I23e13f4073a45b688b1df59d6573625e5a893106